+++ Libraries/SideScroll/Extensions/TypeExtensions.cs
@@ -100,7 +100,21 @@ public static class TypeExtensions
 		{
 			return type.GetElementType();
 		}
-		else if (type.GenericTypeArguments.Length > 0)
+
+		if (type.IsGenericType && type.GetGenericTypeDefinition() == typeof(IEnumerable<>))
+		{
+			return type.GetGenericArguments()[0];
+		}
+
+		Type? enumerableType = type.GetInterfaces()
+			.FirstOrDefault(i => i.IsGenericType && i.GetGenericTypeDefinition() == typeof(IEnumerable<>));
+
+		if (enumerableType != null)
+		{
+			return enumerableType.GetGenericArguments()[0];
+		}
+
+		if (type.GenericTypeArguments.Length > 0)
 		{
 			return type.GenericTypeArguments[0];
 		}
@@ -121,7 +135,7 @@ public static class TypeExtensions
 			.Where(p => p.GetCustomAttribute<HiddenAttribute>() == null)
 			.Where(p => p.GetCustomAttribute<HiddenColumnAttribute>() == null)
 			.Where(p => p.GetIndexParameters().Length == 0)
-			.Where(p => !p.GetAccessors(nonPublic: true)[0].IsStatic)
+			.Where(p => p.GetMethod is { IsPublic: true, IsStatic: false })
 			.OrderBy(p => p.Module.Name)
 			.ThenBy(p => p.MetadataToken)
 			.ToList();
@@ -166,21 +180,27 @@ public static class TypeExtensions
 	/// </summary>
 	public static string GetAssemblyQualifiedShortName(this Type type)
 	{
-		string name;
-		if (type.IsGenericType)
-		{
-			var args = type.GetGenericArguments().Select(a => '[' + a.GetAssemblyQualifiedShortName() + ']');
-			name = $"{type.GetGenericTypeDefinition().FullName}[{string.Join(", ", args)}]";
-		}
-		else
-		{
-			name = type.FullName!;
-		}
+		string name = GetShortFullName(type);
 
 		if (type.Assembly.GetName().Name is { } assemblyName)
 		{
 			name += ", " + assemblyName;
 		}
 		return name;
+
+		static string GetShortFullName(Type t)
+		{
+			if (t.HasElementType)
+			{
+				Type elementType = t.GetElementType()!;
+				return GetShortFullName(elementType) + t.Name[elementType.Name.Length..];
+			}
+			if (t.IsGenericType)
+			{
+				var args = t.GetGenericArguments().Select(a => '[' + a.GetAssemblyQualifiedShortName() + ']');
+				return $"{t.GetGenericTypeDefinition().FullName}[{string.Join(", ", args)}]";
+			}
+			return t.FullName!;
+		}
 	}
 }